# API.yaml
openapi: 3.0.3
info:
  title: SchoolWall API
  description: SchoolWall 项目的后端 API 接口文档。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
...
paths:
...
  /comments:
    post:
      tags:
        - Comments
      summary: 创建评论
      description: 创建新的评论，支持对帖子或其他评论的回复。当回复其他评论时，parentId 应该是被回复评论的 ID，否则 parentId 应是0。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentRequest"
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentDto"
  /comments/me:
    get:
      tags:
        - Comments
      summary: 获取我的评论历史
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Direction"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCommentDto"
  /comments/users/{userId}:
    get:
      tags:
        - Comments
      summary: 获取指定用户的评论
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCommentDto"
  /comments/{commentId}:
    get:
      tags:
        - Comments
      summary: 获取评论详情
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentDto"
    put:
      tags:
        - Comments
      summary: 更新评论
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentUpdateRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentDto"
    delete:
      tags:
        - Comments
      summary: 删除评论
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 删除成功
  /comments/{commentId}/replies:
    get:
      tags:
        - Comments
      summary: 获取评论的回复
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCommentDto"
  /comments/post/{postId}/toplevel:
    get:
      tags:
        - "Comments"
      summary: "获取帖子的所有顶级评论"
      description: "根据帖子ID分页获取所有顶级评论（即直接评论帖子的评论，parent_id为null）。"
      operationId: "getTopLevelCommentsByPostId"
      parameters:
        - name: "postId"
          in: "path"
          description: "帖子ID"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "page"
          in: "query"
          description: "页码，从0开始"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: 0
        - name: "size"
          in: "query"
          description: "每页数量"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: 10
      responses:
        "200":
          description: "成功获取顶级评论列表"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCommentDto"
        "404":
          description: "帖子不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # AdminUserController
...
  /admin/comments/users/{userId}:
    get:
      tags:
        - Admin - Comment Management
      summary: 获取用户的评论
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCommentDto"
...
  /admin/comments/{commentId}:
    delete:
      tags:
        - Admin - Comment Management
      summary: 物理删除评论
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 操作成功
...
  /emotion/comment/{commentId}:
    get:
      summary: 获取或生成评论的情绪分析
      tags:
        - Emotion
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
          description: 评论ID
      responses:
        "200":
          description: 情绪分析结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmotionDto"
        "404":
          description: 评论不存在
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Page:
      name: page
      in: query
      description: 页码 (从0开始)
      schema:
        type: integer
        default: 0
    Size:
      name: size
      in: query
      description: 每页数量
      schema:
        type: integer
        default: 10
    Sort:
      name: sort
      in: query
      description: 排序字段
      schema:
        type: string
    Direction:
      name: direction
      in: query
      description: 排序方向
      schema:
        type: string
        enum: [ASC, DESC]
        default: DESC
  schemas:
...
    CommentRequest:
      type: object
      properties:
        postId:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
        content:
          type: string
    CommentUpdateRequest:
      type: object
      properties:
        content:
          type: string
    CommentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        authorInfo:
          type: object
        postId:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isDeleted:
          type: boolean
...
    PageResponse:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
...
    PageResponseCommentDto:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/CommentDto"
...
    EmotionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 情绪分析记录ID
        postId:
          type: integer
          format: int64
          nullable: true
          description: 帖子ID
        commentId:
          type: integer
          format: int64
          nullable: true
          description: 评论ID
        text:
          type: string
          description: 被分析的文本内容
        sentiment:
          type: string
          description: 情绪标签（如 positive, negative, neutral）
        confidence:
          type: number
          format: double
          description: 情绪置信度
        probabilities:
          type: object
          additionalProperties:
            type: number
            format: double
          description: 各情绪类别的概率分布
        createdAt:
          type: string
          format: date-time
          description: 分析生成时间